# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payments/portal/status/{id}": {
      "get": {
        "operationId": "PaymentsController_getStatusFromPortal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "description": "Payment provider (mtn, airtel, etc). Defaults to mtn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status from provider (portal)"
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/portal/{id}": {
      "get": {
        "operationId": "PaymentsController_findOneFromPortal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment fetched from admin portal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/portal/balance/available": {
      "get": {
        "operationId": "PaymentsController_getBalanceFromPortal",
        "parameters": [
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "description": "Payment provider (mtn, airtel, etc). Defaults to mtn",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet balance for the tenant (portal)"
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments": {
      "get": {
        "operationId": "PaymentsController_findAll",
        "parameters": [
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant ID for multi-tenancy",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID for idempotent request deduplication (recommended)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant ID for multi-tenancy",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/portal": {
      "get": {
        "operationId": "PaymentsController_getFromPortal",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Payments fetched from admin portal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "post": {
        "operationId": "PaymentsController_createFromPortal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created from admin portal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/status/{id}": {
      "get": {
        "operationId": "PaymentsController_getStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "description": "Payment provider (mtn, airtel, etc). Defaults to mtn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant ID for multi-tenancy",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status from provider"
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Payment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant ID for multi-tenancy",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/v1/payments/balance/available": {
      "get": {
        "operationId": "PaymentsController_getBalance",
        "parameters": [
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "description": "Payment provider (mtn, airtel, etc). Defaults to mtn",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant ID for multi-tenancy",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "description": "API key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet balance for the tenant"
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Payment Gateway",
    "description": "API documentation for Payment Gateway",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "x-tenant-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id"
      }
    },
    "schemas": {
...
      "Payment": {
        "type": "object",
        "properties": {}
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "MTN",
              "AIRTEL",
              "ZAMTEL"
            ],
            "example": "MTN",
            "description": "Payment provider (e.g., MTN, AIRTEL, ZAMTEL)"
          },
          "amount": {
            "type": "number",
            "example": 100
          },
          "currency": {
            "type": "string",
            "example": "EUR",
            "default": "EUR"
          },
          "externalId": {
            "type": "string",
            "example": "INV-20260105-001"
          },
          "payer": {
            "type": "string",
            "example": "260765725317"
          },
          "payerMessage": {
            "type": "string",
            "example": "Payment for invoice"
          },
          "payeeNote": {
            "type": "string",
            "example": "Thank you"
          }
        },
        "required": [
          "provider",
          "amount",
          "currency",
          "externalId",
          "payer",
          "payerMessage",
          "payeeNote"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "x-api-key": []
    },
    {
      "x-tenant-id": []
    }
  ]
}